Rework command engine. Way less boilerplate code for new commands

+++ src/main/scala/jk_5/nailed/server/player/NailedPlayer.scala
@@ -187,7 +187,7 @@ class NailedPlayer(private val uuid: UUID, private var name: String) extends Pla
   override def addExperience(level: Int) = entity.addExperience(level)
   override def setExperienceLevel(level: Int) = entity.experienceLevel = level
   override def experienceLevelCap = entity.xpBarCap()
-  override def getExperience = entity.experience
+  override def getExperience = entity.experience.toInt
   override def getExperienceLevel = entity.experienceLevel
 
   override def toString = s"NailedPlayer{uuid=$uuid,name=$name,isOnline=$isOnline,gameMode=$getGameMode,eid=${getEntity.getEntityId}}"
